+++ R/flatten.R
@@ -9,7 +9,10 @@ vecvec_flatten_adj <- function(x) {
 
   i <- 1L
   while (i < length(x)) {
-    if (identical(ptypes[[i]], ptypes[[i + 1L]])) {
+    if (
+      identical(ptypes[[i]], ptypes[[i + 1L]]) && 
+      !(is_altrep(x[[i]]) || is_altrep(x[[i + 1L]]))
+    ) {
       x[[i]] <- c(x[[i]], x[[i + 1L]])
       x <- x[-(i + 1L)]
       ptypes <- ptypes[-(i + 1L)]