+++ R/utils.R
@@ -73,3 +73,10 @@ overlap_indices <- function(a, b) {
     )
   )
 }
+
+# Checks if an object is an ALTREP vector
+is_altrep <- function(x) {
+  internal <- get(".Internal", envir = baseenv())
+  out <- utils::capture.output(internal(inspect(x)))
+  any(grepl("compact|wrapper", out))
+}