+++ NEWS.md
@@ -1,3 +1,5 @@
+# vecvec (development version)
+
 # vecvec 1.0.0
 
 The `vecvec` class has been reimplemented using S7, replacing the previous vctrs