+++ NEWS.md
@@ -1,4 +1,4 @@
-# vecvec (development version)
+# vecvec 1.1.0
 
 ## Improvements